Moved methods for adding content to github from gh to module.
