+++ proteobench/github/gh.py
@@ -4,8 +4,6 @@ from tempfile import TemporaryDirectory
 import pandas as pd
 from git import Repo
 
-import proteobench.modules.dda_quant.module as dda_quant_module
-
 
 def clone_repo_anon(
     clone_dir="K:/pb/",
@@ -15,7 +13,7 @@ def clone_repo_anon(
     return clone_dir
 
 
-def read_json_repo(
+def read_results_json_repo(
     remote_git_repo= "https://github.com/Proteobench/Results_Module2_quant_DDA.git"
 ):
     t_dir = TemporaryDirectory().name
@@ -25,72 +23,8 @@ def read_json_repo(
         remote_git_repo
     )
     fname = os.path.join(t_dir, "results.json")
-    print(fname)
     all_datapoints = pd.read_json(fname)
-    print(all_datapoints)
     return all_datapoints
-
-
-def write_json_local_development(
-    temporary_datapoints
-):  
-    t_dir = TemporaryDirectory().name
-    os.mkdir(t_dir)
-
-    current_datapoint = temporary_datapoints.iloc[-1]
-    current_datapoint["is_temporary"] = False
-    all_datapoints = dda_quant_module.add_current_data_point(None, current_datapoint)
-
-    # TODO write below to logger instead of std.out
-    fname = os.path.join(t_dir, "results.json")
-    print(f"Writing the json to: {fname}")
-
-    f = open(os.path.join(t_dir, "results.json"), "w")
-    
-    all_datapoints.to_json(
-        f,
-        orient="records",
-        indent=2
-    )
-
-    return os.path.join(t_dir, "results.json")
-
-def clone_pr(
-    temporary_datapoints,
-    token,
-    username="Proteobot",
-    remote_git="github.com/Proteobot/Results_Module2_quant_DDA.git",
-    branch_name="new_branch",
-):
-    t_dir = TemporaryDirectory().name
-
-    clone_repo(clone_dir=t_dir, token=token, remote_git=remote_git, username=username)
-    current_datapoint = temporary_datapoints.iloc[-1]
-    current_datapoint["is_temporary"] = False
-    all_datapoints = dda_quant_module.add_current_data_point(None, current_datapoint)
-    branch_name = current_datapoint["id"]
-
-    # do the pd.write_json() here!!!
-    print(os.path.join(t_dir, "results.json"))
-    f = open(os.path.join(t_dir, "results.json"), "w")
-    
-    all_datapoints.to_json(
-        f,
-        orient="records",
-        indent=2
-    )
-
-    f.close()
-    commit_message = "Added new run with id " + branch_name
-
-    pr_github(
-        clone_dir=t_dir,
-        token=token,
-        remote_git=remote_git,
-        username=username,
-        branch_name=branch_name,
-        commit_message=commit_message,
-    )
 
 
 def clone_repo(
