+++ webinterface/pages/TEMPLATE.py
@@ -19,8 +19,6 @@ import streamlit as st
 import streamlit_utils
 from streamlit_extras.let_it_rain import rain
 
-from proteobench.github.gh import clone_pr, write_json_local_development
-
 logger = logging.getLogger(__name__)
 
 ## Different parts of the web application
@@ -225,7 +223,7 @@ class StreamlitUI:
             if submit_pr:
                 st.session_state[SUBMIT] = True
                 if not LOCAL_DEVELOPMENT:
-                    clone_pr(
+                    Module().clone_pr(
                         st.session_state[ALL_DATAPOINTS],
                         st.secrets["gh"]["token"],
                         username="Proteobot",
@@ -233,7 +231,7 @@ class StreamlitUI:
                         branch_name="new_branch",
                     )
                 else:
-                    DDA_QUANT_RESULTS_PATH = write_json_local_development(
+                    DDA_QUANT_RESULTS_PATH = Module().write_json_local_development(
                         st.session_state[ALL_DATAPOINTS]
                     )
         if SUBMIT in st.session_state:
